requirements.txt
apache-airflow==2.5.0
apache-airflow[gcp]
dbt-core==1.3.0
dbt-bigquery
apache-airflow-providers-google==8.9.0
